Created a class for unit tests and added reset_board() to the TicTacToe class.
TicTacToe.reset_board() sets all of the instance variables to an initial
pre-game state.  It's called by the constructor, and will also be called
in the main game loop to reset the board for a new game after a game is
completed.

unittests.py contains a class that inherits from unittest.TestCase to test
that TicTacToe objects behave as they are expected to behave.