Added make_move() and move_possible() methods to TicTacToe class, plus unittests
There are two new methods on the TicTacToe class.  One is move_possible()
which takes a board position as an argument and determines if that
position is free.  The other is make_move() which takes a player and board
position as arguments and makes a move at the specified position for the
player, if the move is possible.  Created unit tests testing these new
methods.