Added a check_game_over() function to TicTacToe and made a unit test for it.
TicTacToe.check_game_over() takes the players' current squares and checks
for matches with the TicTacToe.victories list.  If there is a match, it
returns a tuple (True, score) where score is -1 if the victor was the
human player, or 1 if the computer won.  It also checks for tie games,
returning (True, 0) if the board is full.  If the game is not over, it
returns (False, None).

The scores (-1, 0, 1) will be used to evaluate potential moves in the
recursive method the computer will use to find a non-losing move.