Added do_human_move(), print_board() and play() methods to TicTacToe
TicTacToe.print_board() outputs the board and a table of available moves
to the screen.

TicTacToe.do_human_move() takes user input and attempts to move at the
position specified by the user.

TicTacToe.play() sets up a dummy game where the human player can make
multiple moves until a victory condition is reached.  In the future it will be
modified to play a normal game with human and computer moves.