Added do_computer_turn() method to TicTacToe, with unit test
TicTacToe.do_computer_turn() will try to take the middle square or a
corner for its first move.  For subsequent moves, it will call
TicTacToe.find_good_move(), which will recursively find an optimal move
for the computer.  find_good_move() isn't implemented yet, so it simply
returns a random free space.

TicTacToe.play() was also updated to allow for a real game to be played,
with a very stupid computer opponent.