Patched  tic tac toe game loss with a hack
After playing the tic tac toe game a few more times, I realized I could indeed defeat it. So I created a page to
redirect the player if they won the game so they don't get to enjoy a won game board.

In the javascript source code of the game, when a win is detected the user is immediately redirected to page where
they're are taunted.