Added extensive AI description
I have a list of tuples which contain the winning sets. Currently the AI
is popping all the selections done from the list by the player, and
choosing whichever scenario is most likely to happen. From the three
possible choices--minus any that have been popped--the ai choices the
remaing.

This should always lead to a tie.

NOTE: With that logic, I can definitely shorten the code instead of the
algorithm route I was originally going to take.