Completed Tic-Tac-Toe HTML5/JavaScript Game
HTML was used to layout the game, CSS used to style and set visual
properties of the form. JavaScript is the heart of the game. Player gets
first move, each move is followed by a computer move. The board is
checked to ensure only legal moves are permitted. Each user and computer
each have an uninterrupted turn. The computer will try to win with 1
move, followed by prevent the user from winning with 1 move, followed by
taking 1 of 5 key squares (center and corners), followed by randomly
picking any remaining square. Wins, losses, draws and number of possible
plays are shown at the top of the page. A single button is used to
restart the game should the user feel the need to abort the current
game. Cookies are used to save the current game's information so that
the user can simply reload the page and resume their game.