Creation of start.py, a file to define a simple HTTP web server that is the basis for the Tic-Tac-Toe game.
Minutia about approach: Yes, we could write a standalone program run from the command line. But HTML5 and CSS3 provide a wonderful syntax and behavioral layer. The AI will be cleanly separated from the presentation, so that it can be reused as it's own library. Once more, the HTTP Server will communicate via the client through a simple JSON-over-HTTP, so that if the front end wanted to be swapped out, it could.