Configured Local Time
Cox Media Group is located in the Eastern Time Zone. This time zone has
been used respectively. The Python TimeZone Library has also been
included for better TimeZone handling.

Side Note: Depending on the requirements of the system, one would want
to display timezone information based upon a User's timezone. However,
that is not a requirement or a high-priority feature for this specific
project.