Included CSS to format and 'animate' the table.
The table used for the game is now centered in the screen. It is set to
be wide enough to be easily playable but small enough that it *should*
look okay on a mobile device.

When you hover over a cell, it will highlight the background color in a
blue-ish tint to help indicate selectability.