Stubbed a sample game-board template.
This template currently just displays a simple header, a table which
represents the game-board, and a single icon which represents a User
"game piece".

It imports the following JavaScript and CSS dependencies:

* Bootstrap 3
* Font-Awesome
* jQuery
* AngularJS

At this point, Bootstrap and FontAwesome are the only external
dependencies used. In the future, I expect to use jQuery and AngularJS
to provide a "dynamic" User Experience.