Implemented a disabled state for Game Board Cells.
When a User adds a piece to the board, that cell (or box) is no longer
available to add new pieces. I have labeled this a 'disabled state'.

During a disabled state, the game board cell will show a red background
if the User hovers over it. I may decide later to remove the 'pointer'
cursor in this state as well.