Moved the Game Board to its own template.
The Game Board visualization is going to have quite a bit of its own
functionality and presentation logic. It will also be initialized from
the launch of the page. I've decided to move it to its own template to
simplify the 'global' template.

Date has been included in each Table Cell element to present which
piece belongs in each box.