Stubbed a view to handle piece selections.
This View will be called when a User selects a piece. In theory, the
User will submit a form with a single piece selected. This view will
then determine the game state, assign the new selection if it's
available, execute the computer's next move, save the Game Board, and
redirect the User back to the Game Board View.