Included messaging to display Victory Status.
I've included the human's victory, even though they should never be
able to win. At this point, the computer doesn't take any turns so this
is more-so for testing and future-proofing purposes.

Also, allowing a User to select a new piece has been short-circuited if
the game has already been won.