Implemented check for a Draw.
The game now checks to see if there is a Draw (Tie). This is only done
if *all* of the Game Pieces on the board have been selected and there
are no more Selections available.

A Draw can still be determined at least one move sooner in some cases.
Enhancements can be performed later to improve this behavior.

Fixes #3.