My first commit of code challenge :  Intructions: 1. Fork this repo on GitHub 2. Create a program that can interactively play the game of Tic-Tac-Toe against a human player and never lose. 3. Commit early and often, with good messages. 4. Push your code back to GitHub and send us a pull request.
If you don't want to broadcast your intentions by forking this, feel free to clone it and work locally. Then, send us a tar.gz of your solution, including your .git folder so we can see your commit history.

We are a Django shop, but it is not a requirement that you implement your program as a Django app.