Added AI
This is generating the boards for the available moves. All of the boards are
corrupted; I forgot that Python does everything as a reference, so the fetch
which should be a copy of the board's data is the board's data, and each
call to move() updates all of the boards. That is every board is operating
on the same data.