AI features added
Added the first and second move options based on best practices (avoid the
middle edges, they can only win in 2 ways, center and corner spots can win in
3 ways).

Added a findWinner() which will determine if there is a next move that will win.

Added a mustBlock() which will determine if not taking a spot could result  in
a loss.

Started the __selectBestMove() which will determine the state of the game. While
not finished it will sepect a spot based on
1) First or second moves are in a corner or center spot
2) If a win can be had, take it
3) If a must block situation is found, block the spot
4) (not implemented yet) check to see if two moves ahead will require
   multiple blocks, if so, don't choose an option that leads to this.
5) Pick random from the available spots

Adedd the makeMove() which will make a move on the active board based on the
best move available