Added Game
Added the Game class. This sets up the board, determines the player,
monitors turns, and evaluates the game in progress.

Made changes to supporting modules which came about from testing:

Removed the direct access to the turn in board, renamed turn to 'player'
Added some more spaces in the view for improved readability