Changed logic of space values, refactored get_corner_move and added side logic
Found an edge case where choosing the middle first would allow a player
to win, thus I put the value of a corner higher.

Changed get_corner_move to get_move, this function now has a string that
tells the function if its a side or corner move, this includes a new
get_open_sides method that determines the sides we can choose from.